Do not create static dev nodes in root init
For compatibility reasons kiwi created a set of static device
nodes when initializing a new image root system. With the
presence of devtmpfs this should no longer be needed. In addition
the static dev node setup now also causes problems on filesystems
like btrfs which was the reason to delete this code now.
This Fixes bsc#1087104