Fixup LOADER_LOCATION in sysconfig/bootloader
By default we always set LOADER_LOCATION=mbr which is wrong
if EFI is in use. This patch updates the value to be correct.
It also seems that this variable is only consumed by the
yast2 bootloader module from past days. Thus we consider
it obsolete and on the to be droped list in future releases.
This Fixes #746