Add config-cdroot to description import list
During the prepare step the image description and mandatory
files needed in the create step are copied into the image
root system below the image/ directory. In case of the
optional config-cdroot archive this copy action was missing
which lead to the problem that the archive was not present
if the kiwi system create command is sequence is used.
This Fixes #756