Added system cleanup methods
Some files in the system gets created by services like
systemd or zypper and are meant to be created once on the
target system. However in the image they might be unwanted.
Thus this commit adds convenience methods to delete files
which gets automatically re-created by the services on
startup. Whether or not the methods are used is in the
responsibility of the author of the image descripion.
This Fixes bsc#1098535