Fixed GCE image file name
In former times Google requires the image name to follow
their naming conventions. However that seems to have changed
and it is no longer required to match a certain pattern.
Thus this patch changes the output name of the GCE tar file
to use the same naming schema as KIWI applies to its output
files