Fixed name of checksum file for pxe type
The pxe image build generates among others a checksum
file with the suffix '.md5' This file is read by the
legacy netboot code and is expected to have the same
basename as the image file itself. However if the
compressed attribute is set the image file name is
set to 'image.xz' and the checksum is named 'image.xz.md5'
which is wrong because 'image.md5' is expected. This
patch makes sure the checksum file is always set
to 'image.md5' no matter if the compressed flag is
configured or not