Drop oem-ataraid-scan from schema
Along with the change we pin generateDS version to 2.29.14
because the newer version (2.29.19) creates broken python
syntax for elements with a value list for their content
like it's the case for the packagemanager element. The
tool uses the name k.packagemanager.content as class name
which is invalid for python.