Added support for pxe live boot via AOE
The live ISO should support a network reference. We are using
the Ata Over Ethernet protocol to achieve this. In combination
with pxe boot of the kernel/initrd a live iso can boot from
the network using the following parameter example:

    root=live:AOEINTERFACE:e0.1 rd.kiwi.live.pxe

Export of the live iso file via AOE can be achieved using the
vblade toolkit which needs to be available on the exporting
system and compatible with the live operating system.
This Fixes #796