Fixed overlay of intermediate config files
Some config files e.g etc/hosts needs to be temporary copied
from the buildsystem host to the image root system. This is
done by a custom copy with the .kiwi extension and a symlink
to that file. During the installation process the package
manager either overwrites the file or creates a .rpmnew variant.
In case a .rpmnew variant exists there is code in kiwi which
restores that .rpmnew variant to become the real file. However
that _restore_intermediate_config_rpmnew_variants() method
runs after overlay files has been applied to the system because
it's part of the final cleanup step. In order to preserve
an eventual overlay version of the file the .rpmnew variant
gets only restored if the real file does not exist.
This Fixes #807