Fixes LVM based image creation in OBS
While attempting to create LVM based images under the Open Build
Service I recently ran into some issues related to the fact that
there is no udev running in the chroot environment used to build
kiwi based images.

Two workarounds have been implemented in this patch:

(1) When calling lvcreate, include the `-Zn` option to disable
    the automatic zeroing of the header of the newly created
    LV device; doing so requires that the LV device's /dev
    entry exists immediately after it has been created, but
    in a chroot'd environment udev isn't going to be running
    to automatically populate /dev/<vg_name>/<lv_name> or
    /dev/mapper/<vg_name>-<lv_name>.
    This should be safe to do since the LV is being created
    within a loopback device based partition, which is backed
    by a zero filled file, created by qemu-img.

(2) After creating an LV we need to run `vgscan --mknodes`
    to create the required device nodes under /dev, which
    won't be automatically created since udev isn't running
    in the chroot'd environment.

Unit tests updated to account for additional `-Zn` arguments that
are being passed to `lvcreate` and for additional call that is
being made to `vgscan --mknodes`.

This fixes #824.