Added machine id setup in dracut preparation
In case of a dracut booted image we empty out the systemd
machine-id configuration file to trigger the rebuild of that
information by the dracut boot code at boot time. This allows
for unique systemd identifiers if the same image gets deployed
on different machines. This also makes the script implementations
people put in in config.sh or images.sh to solve this problem
obsolete. This Fixes #843