Allow simple path source in Uri class
This patch is needed as follow up fix for the setup of the
package cache in local repositories. The is_remote method
from the Uri class is used to identify if a repostory source
is remote or local. At that point the initial repository
source was already translated into its components. In case
of a local repository the Uri instance now receives a simple
path and the is_remote method raised with a style error.
This patch allows the Uri class to be more friendly and
initializes a local path as file:/ typed source.
Related to Issue #847