Warn on Modifications to intermediate config files
Some files are taken from the host and managed as intermediate
config files during the build of the image. Changes to those
files during the build run by e.g a script will not become
effective because the file gets restored. With this patch we
detect the modification condition and throw a warning message
such that the author of the image can adapt the description
as suggested in the message. This Fixes #817