Fixed firmware strip and lookup for kiwi initrd
In a kiwi initrd the function baseStripFirmware can be
used to strip down the firmware to the actually used
kernel drivers in that initrd. The code to do this was
broken due to the changes in #64cb5b003140a. This
patch done by Ruediger Oertel <ro@suse.de> fixes the
method to work correctly again.