Better exception handling in oem installer
If an error condition applies in the kiwi dump dracut code
the reaction was to stop the process with a dracut die() call.
If the option 'rd.debug' was set on boot, this lead to a debugging
shell which is good, but in a standard process this lead to a
lock of the machine which is an unfortunate situation. This
patch changes the behavior to always print the error message
as a dialog message box on the primary console and reboot
the system after keypress or timeout. In case of the debug
switch configured the system die()'s as before.
This Fixes #884