Use cow file on persistent grub live loop boot
When using tools like live-grub-stick the live iso as generated
by kiwi will be copied as file on the target device and a grub
loopback setup is created there to boot the live system from file.
In such a case the persistent write setup which tries to create
an extra write partition on the target fails in almost all cases
because the target has no free and unpartitioned space available.
Therefore in case of such a loopback mounted system we create a
cow file(live_system.cow) instead of a partition to setup
persistent writing. The cow file will be created in the same
directory the live iso image file was read from grub
This Fixes #860