Another round of OCI tools refactoring
In order to provide buildah support some of the logic about
temporary directories for OCI images creation needs to be moved
to the dedicated OCI tool class.

While umoci can operate in any directory and this is passed as an
argument, this is not the case for buildah. In buildah workflow
the storage path of work-in-progress images/containers is not
customizable neither the mountpoint of the container rootfs.