Make sure manual pages are part of pypi archive
As consequence of the change in the travis setup to make
use of the native 'pages' and 'pypi' providers from travis the
environment for the sdist target has changed. Within the
doc_travis tox env no manual pages was build. In combination
with the 'pypi' provider setup in .travis.yml which uses
skip_cleanup: true, this caused the pypi archive to miss
any manual pages which lead to a runtime error on install
via pip.