Do not assume package manager is always there
This commit modifies the dump_reload_package_database method
for zypper to not assume rpm binary is always part of the image.

An image could be bootstrapped only without zypper or rpm, in that
case it does not make sense and it is not possible to dump and reload
rpmdb.

Fixes #915 and related to #550