Fixed isolinux fallback path
In case isolinux-config failed or does not exist on the
distribution a fallback path is called. That code hardlinks
the files to the isolinux compiled in standard path. However
due to the move of the grub unicode file for iso images
the path contains a directory. Directories can't be hardlinked
thus this patch uses 'cp -l' instead of the 'ln' command to
create the linked target contents.