Extend spare partition setup
The spare partition could be used to introduce one additional
partition table entry. With this patch the following new type
attributes will be added:

* spare_part_fs="fsname"
* spare_part_mountpoint="/location"
* spare_part_is_last="true|false"

Along with the setup of the partition size the filesystem and
its mountpoint can be specified. If set the contents of the
rootfs at the specified spare location will be synced to that
partition. The spare_part_is_last attribute will place the
spare partition at the end of the disk. Note this attribute
is only available for the simple vmx disk type. This is
related to bsc#1129566