Fixed disk detection from root device
The method lookup_disk_device_from_root assigns the disk device
matching the root device uuid. However in a multipath environment
multiple disk devices matches the same root device. The code to
assign the multipath map in this case was missing in the dracut
code base. This Fixes #954 and Fixes bsc#1126283 and bsc#1126318