Fix suseRemoveYaST to work with set -e
When the config.sh is invoked with set -e then suseRemoveYaST causes the whole
script to fail when no yast packages are present. The problem is that `grep
yast` returns 1 in that case.
=> Explicitly ignore the return value of grep

Furthermore, if no YaST packages were found, we don't want to invoke rpm, as
that would fail too (as it is called without any parameters).