Fixed evaluation of @root size setup
The @root volume setup allows to specify a size for the root
volume in an LVM setup. The size setting was not respected and
the root volume was always resized to the rest space available.
This patch fixes the read_volume_setup_all_free method
on the dracut kiwi library to know if there is root volume
setup and respect its contents. This Fixes #983