Abandon the bumpversion helper
The bumpversion helper script updated the reference file
with changes since the latest date of the existing reference.
However even though the two pieces are in chronological order
that does not mean that the concatenated version of those
is still in chronological order. Thus that helper is deleted
and a new helper rebuild_changelog_reference exists. That
rebuilds the entire reference file from the history and
makes sure it is in chronological order. The tool should
only be called if the chronological order of the total changes
information is violated which under normal conditions
does not happen