Fixed gitlab-ci rpm stage localtime setup
By default the timezone of the container is set to UTC. However
the reference changelog entries are created in the CET timezone.
The current implementation uses dateutil.parser.parse to parse
the reference date/time. That call takes the localtime into account
and responds with a warning:

+ UnknownTimezoneWarning: tzname CET identified but not understood

The current code does not convert the CET timezone into the
localtime timezone (UTC in this case) and creates a timezone naive
datetime object. That object however can't be used to compare
with timezone aware datetime objects which causes a runtime
exception. The qick fix here is to set the localtime to the
timezone used in the reference changes file