Fix derived docker images build
This commit fixes the derived docker images when the base image is a
compressed file. After the refactor in #998 the decompression of the
base image and the skopeo call to import the decompressed image happened
in absolutely independent scopes. NamedTemporaryFile python class by
default deletes the created temporary file when the class instance is
destroyed, this class is used to decompress the base image. Because of
isolated scopes at the time skopeo was called the NamedTemporaryFile
instance was not there anymore. This commit makes sure that the
temporary file instance is already in memory at the time skopeo is
called.