Only execute scripts via bash when they are not user executable
Currently we call scripts directly through bash, which has the unfortunate
disadvantage, that the shebang line is completely ignored.
Now we instead check whether the owner of the script is allowed to execute it
and if yes, we let the OS execute it (which takes the shebang into account) or
otherwise call it through bash.