Disable linkcheck/spellcheck
The spellcheck plugin has not been moved to the sphinx 2.0 API
and there are tons of spelling mistakes reported due to the use
of technical terms that are unknown to spellcheck. Also disabled
the linkcheck because it just stands more in our way than it
reports issues that needs to be fixed.