Fix regression on kiwi-repart dracut module
There was a regression introduced in #1035 (048b5c3e). We mask
systemd-fsck-root.service during the repart, but we are not always
unmasking it. If the repart is not possible the service is masked but
not unmasked, which causes the boot to fail.

With this commit after masking the fsck service we trap EXIT and run the
unmask command on script exit. This way we ensure the unmask command is
always executed.