Fix filesystem labels
If multiple volumes were defined they were not properly labeled. This
commit fixes a miss use of custom_args dictonary by the filesystem
base class. This class was modifying the given custom_args that is
passed by reference, thus modifiyng the custom_args instance of the
caller. This issue was causing to propagate the modified
create_options of a filesystem across all the volumes, causing all
volumes to be labelled as ROOT. With this commit Filesystem class
performs a deepcopy of the custom_args dictionary to limit the scope of
any change inside each Filesystem instance.

Fixes #1044