Stabilize Gitlab rpm stage
The rpm stage in the gitlab CI pipeline runs against a collection
of mirror services. If those mirrors are not available the test
fails. Such failed tests are cumbersome because they don't indicate
an error condition we can/should fix. The openSUSE TW mirrors
showed to be unreachable more often than others which might be
because that target changes their content relatively often. This
patch changes the test strategy to allow the TW rpm test to fail
but still keep the other rpm targets mandatory as they don't show
the connection problems.