Fix memory leak in OCI contianer tests
This commit fixes a memory leak in container_image_oci_test caused
by partially mocking a file opening. A ContainerImageOci instance
tries to open and read two files: the runtime configuration, if
present, and the `/.buildenv`, if present. In order to test the
`/.buildenv` presence file opening is mocked causing the runtime
configuration reading fall in an inifite loop. This commit mocks the
the RuntimeConfig instances to avoid this issue.

Fixes #1110