Fix creation of /etc/sysconfig/bootloader for ISOs
This commit forces the creation of /etc/sysconfig/bootloader
before building the squashfs container. Since bootloader configuration
was generally prepared in a different area of the root-tree after
the creation of the squashfs container this sysconfig file was never
included into the image.

It also makes sure that the sysconfig bootloader INFO message is only
shown when the file is actually written.

Fixes #1112