Make extended attribute check an on demand feature
Support for the xattr module is handled differently at the side
of the distribution vendors. There is xattr(GPL) and pyxattr(MIT)
and both might exist or not depending on the version of the desired
distribution. To deal better with this we made the strict requirement
in kiwi for xattr a distribution specific recommendation in the spec
file and changed the code to load the xattr module as an on demand
plugin. If no xattr module support exists the method
target_supports_extended_attributes() will return False which
effectively result in a warning message and rsync operations to run
without the -X and -A options.