Fixed use of Command calls in destructors
The Command class creates namedtuple to store the result
of the call. That tuple is created inside of the Command
namespace. If called inside of a __del__ destructor which
is often the case in kiwi classes I recently observed an
exception from python saying:

  File "/usr/lib64/python3.6/collections/__init__.py", line 429, in namedtuple
  File "<string>", line 1, in <module>
  ModuleNotFoundError: import of builtins halted; None in sys.modules

I do not understand why this is a problem for python now
and only found the solution in making the command_type variable
a global namedtuple.